+++ tests/testthat/setup.R
@@ -5,7 +5,14 @@ withr::defer({
   unlink(jdbcDriverFolder, recursive = TRUE, force = TRUE)
 }, testthat::teardown_env())
 
-if (getOption("dbms") == "postgresql") {
+
+dbms <- getOption("dbms", default = "sqlite")
+if (dbms == "sqlite") {
+  connectionDetails <- Eunomia::getEunomiaConnectionDetails()
+  cdmDatabaseSchema <- "main"
+  cdmVersion <- 5
+}
+if (dbms == "postgresql") {
   DatabaseConnector::downloadJdbcDrivers("postgresql", pathToDriver = jdbcDriverFolder)
   connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                                user = Sys.getenv("CDM5_POSTGRESQL_USER"),
@@ -16,7 +23,18 @@ if (getOption("dbms") == "postgresql") {
   cdmDatabaseSchema <- Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA")
   cdmVersion <- 5
 }
-if (getOption("dbms") == "sql server") {
+if (dbms == "redshift") {
+  DatabaseConnector::downloadJdbcDrivers("redshift", pathToDriver = jdbcDriverFolder)
+  connectionDetails <- createConnectionDetails(dbms = "redshift",
+                                               user = Sys.getenv("CDM5_REDSHIFT_USER"),
+                                               password = URLdecode(Sys.getenv("CDM5_REDSHIFT_PASSWORD")),
+                                               server = Sys.getenv("CDM5_REDSHIFT_SERVER"),
+                                               pathToDriver = jdbcDriverFolder)
+
+  cdmDatabaseSchema <- Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA")
+  cdmVersion <- 5
+}
+if (dbms == "sql server") {
   DatabaseConnector::downloadJdbcDrivers("sql server", pathToDriver = jdbcDriverFolder)
   connectionDetails <- createConnectionDetails(dbms = "sql server",
                                                user = Sys.getenv("CDM5_SQL_SERVER_USER"),
@@ -26,7 +44,7 @@ if (getOption("dbms") == "sql server") {
   cdmDatabaseSchema <- Sys.getenv("CDM5_SQL_SERVER_CDM_SCHEMA")
   cdmVersion <- 5
 }
-if (getOption("dbms") == "oracle") {
+if (dbms == "oracle") {
   DatabaseConnector::downloadJdbcDrivers("oracle", pathToDriver = jdbcDriverFolder)
   connectionDetails <- createConnectionDetails(dbms = "oracle",
                                                user = Sys.getenv("CDM5_ORACLE_USER"),
@@ -34,6 +52,10 @@ if (getOption("dbms") == "oracle") {
                                                server = Sys.getenv("CDM5_ORACLE_SERVER"),
                                                pathToDriver = jdbcDriverFolder)
   cdmDatabaseSchema <- Sys.getenv("CDM5_ORACLE_CDM_SCHEMA")
-  oracleTempSchema <- Sys.getenv("CDM5_ORACLE_OHDSI_SCHEMA")
+
+  # Restore temp schema setting after tests complete
+  oldTempSchema <- getOption("sqlRenderTempEmulationSchema")
+  withr::defer(options("sqlRenderTempEmulationSchema" = oldTempSchema), testthat::teardown_env())
+  options("sqlRenderTempEmulationSchema" = Sys.getenv("CDM5_ORACLE_OHDSI_SCHEMA"))
   cdmVersion <- 5
 }