+++ NEWS.md
@@ -1,3 +1,21 @@
+SelfControlledCohort 1.6.0
+==========================
+
+Changes: 
+
+- Removed support for older versions of DatabaseConnector (< 4.0.0)
+- Deprecated usage of `oracleTempSchema` in favour of `tempEmulationSchema`
+- Allow calls to run with a DatabaseConnector `connection` object 
+- No longer supports undocumented use of `connectionDetails$conn` for storing connections
+  when calling `runSelfControlledCohort`
+- Added function computeSccRiskWindows to allow computation of exposure
+  risk windows without having to run full SCC analysis
+- Storage of results (before rate ratios are computed) now in a results
+  table.
+- Storage of time at risk results no longer merged with estimates
+- Added tests on sqlite for faster unit testing while working on development
+- Added testing on redshift
+
 SelfControlledCohort 1.5.1
 ==========================